+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
 ## [18.0.0] - 2022-12-18
 ### Added
 - [#1898] Keyboard shortcut to remove tags (DEL, SHIFT+DEL)